Changed tests to avoid runtimewarnings without needing to suppress them.

+++ pyxem/generators/variance_generator.py
@@ -125,7 +125,7 @@ class VarianceGenerator():
         """
         im = self.signal.T
         mean_im = im.mean((0, 1))
-        meansq_im = Signal2D(np.square(im.data.astype(np.uint16))).mean((0, 1))
+        meansq_im = Signal2D(np.square(im.data)).mean((0, 1))
         normvar = (meansq_im.data / np.square(mean_im.data)) - 1.
         var_im = Signal2D(normvar)
         corr_var_array = normvar - (np.divide(dqe, mean_im.data))
